Atualização de 22Jan26

+++ src/main/java/br/com/fiap/bo/ServicoBO.java
@@ -1,12 +1,18 @@
 package br.com.fiap.bo;
 
+import br.com.fiap.dao.ProjetoDAO;
 import br.com.fiap.dao.ServicoDAO;
+import br.com.fiap.dao.UsuarioDAO;
+import br.com.fiap.to.ProjetoTO;
 import br.com.fiap.to.ServicoTO;
+import br.com.fiap.to.UsuarioTO;
 
 import java.util.ArrayList;
 
 public class ServicoBO {
     private ServicoDAO servicoDAO = new ServicoDAO();
+    private ProjetoDAO projetoDAO = new ProjetoDAO();
+    private UsuarioDAO usuarioDAO = new UsuarioDAO();
 
     public ArrayList<ServicoTO> findAll() {
         servicoDAO = new ServicoDAO();
@@ -20,8 +26,23 @@ public class ServicoBO {
         return servicoDAO.findById(id);
     }
 
-    public ServicoTO save(ServicoTO servico) {
+    public ServicoTO save(ServicoTO servico) throws Exception {
         servicoDAO = new ServicoDAO();
+        projetoDAO = new ProjetoDAO();
+        usuarioDAO = new UsuarioDAO();
+
+        ProjetoTO servicoProjeto = projetoDAO.findById(servico.getProjeto().getId());
+        if(servicoProjeto == null) {
+            throw new Exception("Não existe um projeto com o id informado");
+        }
+
+        UsuarioTO servicoUsuario = usuarioDAO.findById(servico.getUsuario().getId());
+        if(servicoUsuario == null) {
+            throw new Exception("Não existe um usuario com o id informado");
+        }
+
+        servico.setProjeto(servicoProjeto);
+        servico.setUsuario(servicoUsuario);
 
         return servicoDAO.save(servico);
     }
@@ -29,12 +50,30 @@ public class ServicoBO {
     public boolean delete(long id) {
         servicoDAO = new ServicoDAO();
 
+        if(servicoDAO.findById(id) == null) {
+            return false;
+        }
+
         return servicoDAO.delete(id);
     }
 
-    public ServicoTO update(ServicoTO servico) {
+    public ServicoTO update(ServicoTO servico) throws Exception {
         servicoDAO = new ServicoDAO();
+        projetoDAO = new ProjetoDAO();
+        usuarioDAO = new UsuarioDAO();
 
+        ProjetoTO servicoProjeto = projetoDAO.findById(servico.getProjeto().getId());
+        if(servicoProjeto == null) {
+            throw new Exception("Não existe um projeto com o id informado");
+        }
+
+        UsuarioTO servicoUsuario = usuarioDAO.findById(servico.getUsuario().getId());
+        if(servicoUsuario == null) {
+            throw new Exception("Não existe um usuario com o id informado");
+        }
+
+        servico.setProjeto(servicoProjeto);
+        servico.setUsuario(servicoUsuario);
         return servicoDAO.update(servico);
     }
 }
